+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 # Stream Engine
 
-# Development Release 1.5.0 2018-03-22
+# Development Release 1.5.0 2018-03-26
 
 Issue #13264 - Removed the unused depth variables from NetCDF list